delete images when editing or deleting a product

+++ controller/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require('express-validator/check')
 const Product = require('../models/product')
+const fileHelper = require('../utils/file')
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -145,6 +146,7 @@ exports.postEditProduct = async (req, res, next) => {
       product.price = updatedPrice
       product.description = updatedDescription
       if (image) {
+        fileHelper.deleteFile(product.imageUrl)
         product.imageUrl = image.path
       }
 
@@ -162,7 +164,15 @@ exports.postEditProduct = async (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   const prodId = req.body.productId
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found.'))
+      }
+
+      fileHelper.deleteFile(product.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log('Product Deleted')
       res.redirect('/admin/products')
